requirements.txt
numpy~=1.26.4
pillow~=10.2.0